+++ CHANGELOG.rst
@@ -14,7 +14,10 @@ Changelog
 * Initial implementation.
 
 
-0.0.2 (2024-03-24)
+0.0.2 (2024-03-26)
 ------------------
 
 * Added sorted_query method.
+* Added __len__ and __contains__ methods.
+* Added clear method.
+* Fixed issue with not unique heap keys